+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Text::modify()`, `RichText::push_span()`, and `RichText::modify()` to allow for easier mutation of the underlying values of `Text` and `RichText`
 - `Transform::get_safe_scale` and `Transform::transform_vector`
 - Documented all `VecN` functions
+- Font fallback
+- More granular image features to Cargo.toml
 
 ### Changed
 - `Renderer::create_window_ext` now takes either `WindowProperties` or `AdvancedWindowProperties`
@@ -22,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When using `ViewMode::Stretch`, outlines don't stretch unevenly anymore
 - Cloning `Text` and `RichText` is now cheap and reference-counted
 - Moved `WindowId` and `WindowProperties` from `types` to `window`
+- README update
+- Lighten up dependencies, remove `thiserror` and change default enabled features
+- Boids example now relies on `fastrand` instead of `rand`
 
 ### Removed
 - `RenderSurface::round_rect`, in favor of having corner radius be part of the stateful style
@@ -30,3 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Font::atlas()` and `Font::atlas_mut()` can no longer panic
 - `Canvas::read()` and `Canvas::write()` can no longer panic
 - Wrong/inconsistent documentation
+- Canvases no longer create a new bind group per composite per frame
+- Recursive canvases no longer create a new texture every frame
+- Setting the texture to `A` then `None` then `A` again no longer creates a new batch